third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/viewport-width-subresource.https-expected.txt

This is a testharness.js-based test.
FAIL viewport-width-subresource assert_equals: expected "503" but got "FAIL"
Harness: the test ran to completion.
